+++ sparkle/solver/solver_cli.py
@@ -2,6 +2,7 @@
 # -*- coding: UTF-8 -*-
 """Run a solver, read/write to performance dataframe."""
 
+import sys
 from filelock import FileLock
 import argparse
 from pathlib import Path
@@ -13,9 +14,11 @@ from runrunner import Runner
 from sparkle.solver import Solver
 from sparkle.types import resolve_objective
 from sparkle.structures import PerformanceDataFrame
+from sparkle.tools.solver_wrapper_parsing import parse_commandline_dict
 
 
-if __name__ == "__main__":
+def main(argv: list[str]) -> None:
+    """Main function of the command."""
     # Define command line arguments
     parser = argparse.ArgumentParser()
     parser.add_argument(
@@ -51,7 +54,8 @@ if __name__ == "__main__":
     )
     parser.add_argument(
         "--configuration",
-        type=dict,
+        type=str,
+        nargs="+",
         required=False,
         help="configuration for the solver",
     )
@@ -60,14 +64,20 @@ if __name__ == "__main__":
         "--seed",
         type=int,
         required=False,
-        help="seed to use for the solver. If not provided, read from "
-        "the PerformanceDataFrame or generate one.",
+        help="seed to use for the solver. If not provided, generates one.",
     )
     parser.add_argument(
         "--cutoff-time",
         type=int,
         required=False,
         help="the cutoff time for the solver.",
+    )
+    parser.add_argument(
+        "--objectives",
+        type=str,
+        required=False,
+        nargs="+",
+        help="The objectives to evaluate to Solver on. If not provided, read from the PerformanceDataFrame.",
     )
     parser.add_argument(
         "--target-objective",
@@ -82,11 +92,11 @@ if __name__ == "__main__":
         nargs="+",
         help="If given, will ignore any given configurations, and try to"
         " determine the best found configurations over the given "
-        "instances. Defaults to the first objective given in the "
-        " objectives argument or the one given by the dataframe "
-        " to use to determine the best configuration.",
+        "instances. Uses the 'target-objective' given in the arguments"
+        " or the first one given by the dataframe to determine the best"
+        "configuration.",
     )
-    args = parser.parse_args()
+    args = parser.parse_args(argv)
     # Process command line arguments
     log_dir = args.log_dir
     print(f"Running Solver and read/writing results with {args.performance_dataframe}")
@@ -106,28 +116,48 @@ if __name__ == "__main__":
         run_instances = str(instance_path)
 
     solver = Solver(args.solver)
+    # By default, run the default configuration
+    config_id = PerformanceDataFrame.default_configuration
+    configuration = None
+    # If no seed is provided by CLI, generate one
+    seed = args.seed if args.seed else random.randint(0, 2**32 - 1)
+    # Parse the provided objectives if present
+    objectives = (
+        [resolve_objective(o) for o in args.objectives] if args.objectives else None
+    )
 
-    if args.configuration:
-        configuration = args.configuration
-        config_id = configuration["config_id"]
-    elif args.configuration_id or args.best_configuration_instances:  # Read
+    if args.configuration:  # Configuration provided, override
+        configuration = parse_commandline_dict(args.configuration)
+        config_id = configuration["configuration_id"]
+    elif (
+        args.configuration_id or args.best_configuration_instances or not objectives
+    ):  # Read from PerformanceDataFrame, can be slow
         # Desyncronize from other possible jobs writing to the same file
+        print(
+            "Reading from Performance DataFrame.. "
+            f"[{'configuration' if (args.configuration_id or args.best_configuration_instances) else ''} "
+            f"{'objectives' if not objectives else ''}]"
+        )
         time.sleep(random.random() * 10)
         lock = FileLock(f"{args.performance_dataframe}.lock")  # Lock the file
         with lock.acquire(timeout=600):
             performance_dataframe = PerformanceDataFrame(args.performance_dataframe)
 
-        objectives = performance_dataframe.objectives
-        # Filter out possible errors, shouldn't occur
-        objectives = [o for o in objectives if o is not None]
+        if not objectives:
+            objectives = performance_dataframe.objectives
+
         if args.best_configuration_instances:  # Determine best configuration
             best_configuration_instances: list[str] = args.best_configuration_instances
             # Get the unique instance names
             best_configuration_instances = list(
                 set([Path(instance).stem for instance in best_configuration_instances])
             )
-            target_objective = resolve_objective(args.target_objective)
-            config_id, value = performance_dataframe.best_configuration(
+            target_objective = (
+                resolve_objective(args.target_objective)
+                if args.target_objective
+                else objectives[0]
+            )
+            config_id, _ = performance_dataframe.best_configuration(
                 solver=str(args.solver),
                 objective=target_objective,
                 instances=best_configuration_instances,
@@ -136,16 +166,13 @@ if __name__ == "__main__":
                 str(args.solver), config_id
             )
 
-        elif args.configuration_id:  # Read from DF the ID
+        elif (
+            args.configuration_id
+        ):  # Read from PerformanceDataFrame the configuration using the ID
             config_id = args.configuration_id
             configuration = performance_dataframe.get_full_configuration(
                 str(args.solver), config_id
             )
-
-        seed = args.seed
-        # If no seed is provided and no seed can be read, generate one
-        if not isinstance(seed, int):
-            seed = random.randint(0, 2**32 - 1)
 
     print(f"Running Solver {solver} on instance {instance_name} with seed {seed}..")
     solver_output = solver.run(
@@ -167,15 +194,23 @@ if __name__ == "__main__":
         PerformanceDataFrame.column_value,
         PerformanceDataFrame.column_seed,
     ]
-    objective_values = [
-        f"{objective.name}: {solver_output[objective.name]}" for objective in objectives
-    ]
-    print(f"Appending value objective values: {', '.join(objective_values)}")
+
     print(f"For Solver/config: {solver}/{config_id}")
-    print(f"For index: Instance {instance_name}, Run {args.run_index}")
+    print(f"For index: Instance {instance_name}, Run {args.run_index}, Seed {seed}")
+    print("Appending the following objective values:")  # {', '.join(objective_values)}")
+    for objective in objectives:
+        print(
+            f"{objective.name}, {instance_name}, {args.run_index} | {args.solver}, {config_id}: {solver_output[objective.name]}"
+        )
 
     # Desyncronize from other possible jobs writing to the same file
-    time.sleep(random.random() * 10)
+    time.sleep(random.random() * 100)
+    # TESTLOG
+    # lock = FileLock("test.log.lock")
+    # with lock.acquire(timeout=600):
+    #     with Path("test.log").open("a") as f:
+    #         for objective in objectives:
+    #             f.write(f"{objective.name}, {instance_name}, {args.run_index} | {solver} {config_id}: {solver_output[objective.name]}, {seed}\n")
 
     # Now that we have all the results, we can add them to the performance dataframe
     lock = FileLock(f"{args.performance_dataframe}.lock")  # Lock the file
@@ -189,5 +224,9 @@ if __name__ == "__main__":
             objective=[o.name for o in objectives],
             run=run_index,
             solver_fields=solver_fields,
-            append_write_csv=True,
+            append_write_csv=True,  # We do not have to save the PDF here, thanks to this argument
         )
+
+
+if __name__ == "__main__":
+    main(sys.argv[1:])